Local main had diverged from origin/main (23 local-only security-fix commits,
39 remote-only dependency/CI/test commits). All 23 local commits were verified
unique via `git cherry` and are preserved in this merge. Pre-merge local state
is retained on branch backup/main-pre-pull-20260926-151409.

Conflicts resolved as unions of both sides, never by discarding one:

- api-gateway/gateway_routes.py: local request-body streaming (budgeted via
  BodyLimitMiddleware) + remote response streaming with
  MAX_RESPONSE_BODY_SIZE cap, Connection-nominated header exclusion, and
  raw-header copying. Kept local's POST/PUT/PATCH gate (remote sent an empty
  body on GET/DELETE) and local's rate_limiter call.
- api-gateway/middleware.py: kept local's decompressed-byte budget,
  decompression-bomb guard, 413 handling and StreamingResponse limit; adopted
  remote's exact raw-header preservation. Excluded content-length since the
  body is re-counted as it streams.
- admin-service + streaming-service settings: kept all local secret/URL
  validation and added remote's whitespace-only guard, extra default secret,
  and the PLAYBACK_URL_SIGNING_SECRET check (field verified to exist).
- auth-service conftest: kept local sys.path bootstrap (needed for the
  per-service `app` package + SDK) with remote's black-formatted imports.
  Re-added `import sys` / `from pathlib import Path`, which the auto-merge had
  silently dropped while keeping the block that uses them.
- billing-service test_repositories: adopted remote's PostgreSQL/
  testcontainers fixture; local's `tables=[...]` list was a SQLite-only JSONB
  workaround, now obsolete since CI supplies postgres:15.
- helm values.yaml: took remote's JSON-list CORS_ALLOWED_ORIGINS (pydantic
  parses list[str] from JSON; the bare string is a startup failure) and kept
  service.port 80 (templates use targetPort by name; 8000 was a docker-compose
  host-port carry-over, and 8000 only matches api-gateway).
- ci-cd.yml: pinned actions/upload-artifact to the real v7 SHA
  043fb46d1a93c77aae656e7c1c64a875d1fc6a0a, preserving the #809 supply-chain
  invariant that rejects mutable action refs.
- poetry.lock: kept the auto-merged union (remote's upgrades plus local's
  pylint 4.0.8) and regenerated the content-hash; `poetry check --lock` passes.

Also updated api-gateway/tests/test_body_streaming.py to mock `client.stream`
instead of `client.request`, matching the merged response-streaming path.

Verification: all 15 service suites pass (1447 tests, exit 0);
.github/scripts/verify-supply-chain.py passes; no conflict markers remain.

+++ services/creators-service/app/schemas/creator.py
@@ -1,6 +1,7 @@
 """Pydantic v2 schemas for the Creators service."""
 
 from datetime import datetime
+from decimal import Decimal
 from uuid import UUID
 from typing import Optional
 
@@ -40,7 +41,7 @@ class CreatorAccountResponse(BaseModel):
 
 # ----------------------------------------------------------------- EffectiveFloor
 class EffectiveFloorCreate(BaseModel):
-    per_minute_amount: float = Field(..., ge=0)
+    per_minute_amount: Decimal = Field(..., ge=0, max_digits=20, decimal_places=8)
     currency: str = Field("USD", max_length=8)
     reason: Optional[str] = Field(None, max_length=500)
 
@@ -48,7 +49,7 @@ class EffectiveFloorCreate(BaseModel):
 class EffectiveFloorResponse(BaseModel):
     id: UUID
     creator_id: UUID
-    per_minute_amount: float
+    per_minute_amount: Decimal
     currency: str
     effective_from: datetime
     last_adjusted_at: Optional[datetime]
